[Exchanges] improve exchange ui config and fix issues
