+++ octobot_trading/exchanges/config/ccxt_exchange_ui_settings.pxd
@@ -15,4 +15,4 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-cdef class CCXTExchangeConfig
+cdef void initialize_experimental_exchange_settings(object exchange, dict inputs)
