+++ octobot_trading/exchanges/implementations/spot_ccxt_exchange.py
@@ -26,7 +26,7 @@ from octobot_commons import number_util
 
 import octobot_trading.enums as enums
 import octobot_trading.errors as errors
-from octobot_trading.exchanges.config import ccxt_exchange_settings
+from octobot_trading.exchanges.config import ccxt_exchange_settings, ccxt_exchange_ui_settings
 import octobot_trading.exchanges.types as exchanges_types
 import octobot_trading.exchanges.connectors as exchange_connectors
 import octobot_trading.personal_data as personal_data
@@ -58,7 +58,7 @@ class SpotCCXTExchange(exchanges_types.SpotExchange):
         """
         Called at constructor, should define all the exchange's user inputs.
         """
-        ccxt_exchange_settings.initialize_experimental_exchange_settings(cls, inputs)
+        ccxt_exchange_ui_settings.initialize_experimental_exchange_settings(cls, inputs)
 
 
     @classmethod
